Add form submit for forms displayed in chat
- Fix form submit to work for forms displayed in the chat box
- Close #1797
